Created UI reducer
